requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
asgiref==3.3.4
attrs==20.2.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
comtypes==1.1.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.2
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.1.0
idna==2.9
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
JayDeBeApi==1.2.3
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.15.1
JPype1==1.0.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-server==1.2.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mne==0.20.5
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
nose==1.3.7
notebook==6.1.3
oauthlib==3.1.0
openpyxl==3.0.4
packaging==20.4
pandas
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.6.1
PyMySQL==1.0.2
pyparsing==2.4.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2019.3
pyzmq==19.0.2
requests==2.23.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==5.0.4
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
wfdb==3.0.1
zipp==3.1.0
